+++ src/sn/AnalyzeTagArg.ts
@@ -65,8 +65,16 @@ export class AnalyzeTagArg {
 			const {key, val, val2, literal} = groups!;
 			if (literal) {
 				if (literal.slice(-1) === '=') {
-					const {ln: k_ln, ch: k_ch} = this.#idx2LnCol(lenNm, ln, ch, args, index);
-					hRng[literal.slice(0, -1)] = {k_ln, k_ch, v_ln: ln, v_ch: 1+lenNm +ch +literal.length +1, v_len: 0};
+					const lenVnm = literal.length -1;
+					const {ch: k_ch} = this.#idx2LnCol(lenNm, ln, ch, args, index +lenVnm);
+					hRng[literal.slice(0, -1)] = {
+						k_ln: ln,
+						k_ch: k_ch -lenVnm,
+						v_ln: ln,
+						v_ch: k_ch +1,
+					//	v_ch: ch +1+lenNm +literal.length +1,
+						v_len: 0,
+					};
 				}
 				continue;
 			}